+++ gDRutils/inst/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Change to v.1.3.3
+- Move json validation/convertion logic from gDRelastic
+
+# Change to v.1.3.2
+- Standardize MAE using default gDR identifiers
+
+# Change to v.1.3.1
+- Remove adding integer identifiers at the end of colnames/rownames
+
 # Change to v.1.3.0
 - Release